Adding delete icon, delete functionality from database

+++ src/components/EventsList.vue
@@ -9,7 +9,9 @@
         >
           <!-- <router-link :to="`/events/${user.userId}`"> -->
           <router-link :to="`/events`"> 
-            <strong>{{ event.name }}</strong> ({{ event.count }} contacts)<br>
+            <img class="delete-logo" @click="handleDelete" width="15px" src="../assets/delete-icon.png">
+            <strong>{{ event.name }}</strong> ({{ event.count }} contacts)
+            <br>
           </router-link>
           {{ event.description }}<br>&nbsp;
         </li>
@@ -26,7 +28,7 @@ export default {
     return {
       events: null,
       error: null, 
-      userId: this.user.userId
+      userId: this.user.userId,
     };
   },
   created() {
@@ -41,16 +43,22 @@ export default {
   },
   props: ['user'],
   methods: {
-    deleteEvent(event) {
-      const id = event.eventId;
-      // remove from server
-      return deleteEvent(this.eventId)
-        .then(() => {
+    handleDelete(event) {
+      const confirmDelete = confirm('Are you sure you want to delete this event and all associated contacts?');
+      if(confirmDelete === true) {
+        console.log('Clicked confirm');
+        const id = event.eventId;
+        // remove from server
+        return deleteEvent(this.eventId)
+          .then(() => {
           // remove from current list of events
-          const index = this.events.findIndex(event => event.eventId === id);
-          if(index === -1) return;
-          this.events.splice(index, 1);
-        });
+            const index = this.events.findIndex(event => event.eventId === id);
+            if(index === -1) return;
+            this.events.splice(index, 1);
+          });
+      } else {
+        console.log('Clicked cancel');
+      }
     }
   }
 };
@@ -72,6 +80,9 @@ ul {
 }
 li {
   max-width: 375px;
+}
+.delete-logo {
+  margin-right: 5px;
 }
